% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{AddTag}
\alias{AddTag}
\title{Add Umap Tag}
\usage{
AddTag(
  p,
  aspect.ratio = 1,
  strip.font.size = 8,
  legend.position = "right",
  legend.font.size = 5,
  line.length = 3,
  line.color = "black",
  line.lab.size = 2
)
}
\arguments{
\item{p}{seurat plot}

\item{aspect.ratio}{default is 1}

\item{strip.font.size}{default is 12}

\item{legend.position}{default is right, set none if not needed}

\item{legend.font.size}{default is 8}

\item{line.length}{default is 3}

\item{line.color}{default is black}

\item{line.lab.size}{default is 3}
}
\value{
Returns plot object
}
\description{
Add umap tag on the topbottom
}
\concept{utility}
