% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debackground.R
\name{background.removal}
\alias{background.removal}
\title{Remove background RNA using SoupX}
\usage{
background.removal(raw, filtered, outdir, project)
}
\arguments{
\item{raw}{cellranger output raw_feature_bc_matrix folder}

\item{filtered}{cellranger output filtered_feature_bc_matrix folder}

\item{outdir}{output folder}

\item{project}{project name}
}
\description{
tutorial: https://rawcdn.githack.com/constantAmateur/SoupX/204b602418df12e9fdb4b68775a8b486c6504fe4/inst/doc/pbmcTutorial.html
output is the count matrix
}
