% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{basic_analysis,Input-method}
\alias{basic_analysis,Input-method}
\title{Basic Analysis Using Seurat}
\usage{
\S4method{basic_analysis}{Input}(object)
}
\description{
from cellranger matrix to cluster, including normalization, scaling,
dimension reduction, clustering
}
