% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/communication.R
\name{CellCommunication}
\alias{CellCommunication}
\alias{CellCommunication.CellChat}
\title{Cell Communication}
\usage{
CellCommunication(infile, outdir, used = "CellChat")

\method{CellCommunication}{CellChat}(infile, outdir, species = NULL)
}
\arguments{
\item{infile}{Input file path, supported format is h5ad, loom, rds}

\item{outdir}{Output path}

\item{used}{Methods for anlysis: CellChat}

\item{species}{Species, only support human or mouse}
}
\value{
Return a combined object
}
\description{
Cell Communication
}
\section{CellChat}{

\itemize{
\item source code: \url{https://github.com/wu-yc/scMetabolism}
\item quick start: \url{https://github.com/wu-yc/scMetabolism}
}
}

\concept{Cell Communication}
\concept{Cell to Cell Communication}
