% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cellmarkers}
\alias{cellmarkers}
\title{World Health Organization TB data}
\format{
\subsection{\code{cellmarker}}{

A data frame with 7,240 rows and 60 columns:
\describe{
\item{species}{species name, Human or Mouse}
\item{tissue_class}{tissue name}
\item{cell_name}{cell name}
\item{Symbol}{gene symbol}
\item{technology_seq}{sequencing method}
\item{marker_source}{Experiment, Review, ...}
}
}
}
\source{
\url{http://bio-bigdata.hrbmu.edu.cn/CellMarker/index.html}
}
\usage{
cellmarkers
}
\description{
A subset of data from the World Health Organization Global Tuberculosis
Report ...
}
\keyword{datasets}
