% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{clustering}
\alias{clustering}
\alias{renameClusterPlotMarkers}
\title{Clustering on seurat object}
\usage{
clustering(
  input,
  outdir,
  project,
  dims,
  nfeatures = 2000,
  plot.features = c("nFeature_RNA", "percent.mt", "percent.rb"),
  reduction = "pca",
  k = 20,
  resolution = 2
)

renameClusterPlotMarkers(
  input,
  outdir,
  project,
  groupfile,
  markerfile,
  key = "celltype_fine"
)
}
\arguments{
\item{input}{Input seurat object}

\item{outdir}{Output folder}

\item{project}{Project name}

\item{nfeatures}{Number of variable features to used}

\item{plot.features}{Features to plot on UMAP, default is c("nFeature_RNA", "percent.mt", "percent.rb")}

\item{reduction}{Reduction method, current support is pca, harmony, iNMF (liger)}

\item{resolution}{Resolution to use for clustering}

\item{groupfile}{Group file, "cluster,cellType"}

\item{markerfile}{Marker file, "cellType,gene"}

\item{key}{Cell type name, default is "celltype_fine"}

\item{dim}{Dimensions to use for clustering}
}
\description{
Clustering on seurat object

Rename cluster and plot cluster markers
}
