% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ConvertH5adToSeurat}
\alias{ConvertH5adToSeurat}
\title{Read H5ad to Seurat Object}
\usage{
ConvertH5adToSeurat(infile, outfile)
}
\arguments{
\item{infile}{input is h5ad file}
}
\value{
Returns Seurat objects
}
\description{
Merge data included in the csvfile to one list
}
\concept{utility}
