% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{determineOptimalDims}
\alias{determineOptimalDims}
\title{Determine Optimize Dimensionality}
\usage{
determineOptimalDims(data)
}
\arguments{
\item{input}{elbow plot data}
}
\value{
Returns optimal dims
}
\description{
determine optiomal dimensionality
}
