% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{DeterminePCS}
\alias{DeterminePCS}
\title{Determine PCS by stdev}
\usage{
DeterminePCS(obj.seu)
}
\arguments{
\item{obj.seu}{seurat object}
}
\value{
Returns optimal dims
}
\description{
determine optimal pcs
}
