% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{Input-class}
\alias{Input-class}
\title{Input class for basic analysis}
\description{
Input class for basic analysis
}
\section{Slots}{

\describe{
\item{\code{indir}}{cellranger count matrix directory}

\item{\code{outdir}}{output directory}

\item{\code{steps}}{list of steps(default: 1,2,3,4,5,8):
1-QualityControl, 2-Normalization, 3-FeatureSelection,
4-Scaling, 5-dimensionalReduction, 6-RemoveDoublet,
7-DifferentialExpression, 8-AutoCellType}

\item{\code{maxrna}}{qc thresold: max rna counts}

\item{\code{minrna}}{qc thresold: min rna counts}

\item{\code{mincell}}{qc thresold: min cell counts}

\item{\code{maxmt}}{qc thresold: max mt percentage, default: 5}

\item{\code{project}}{project name}
}}

