% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{MergeFileData}
\alias{MergeFileData}
\title{Merge Data to List of Seurat Object}
\usage{
MergeFileData(csv, outdir, mincell, minrna, maxrna, maxmt, genecol = 2)
}
\arguments{
\item{csv}{The header must be "project,path"}
}
\value{
Returns List consisted of multiple Seurat objects
}
\description{
Merge data included in the csvfile to one list
}
\concept{utility}
