% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/metabolism.R
\name{MetabolicAnalysis}
\alias{MetabolicAnalysis}
\alias{MetabolicAnalysis.scMetabolism}
\title{Metabolic Analysis}
\usage{
MetabolicAnalysis(csv, outdir, used = "scMetabolism")

\method{MetabolicAnalysis}{scMetabolism}(infile, outdir, pathways = NULL)
}
\arguments{
\item{outdir}{Output path}

\item{used}{Methods for anlysis: scMetabolism, scFEA}

\item{infile}{Input file path, supported format is h5ad, loom, rds}

\item{object}{The header is "project,path", the path must included matrix.mtx, genes.tsv, bardcodes.tsv}

\item{pathway}{Interested Pathway, such as "Glycolysis / Gluconeogenesis"}
}
\value{
Return a combined object
}
\description{
Metabolic Analysis
}
\section{scMetabolism}{

\itemize{
\item source code: \url{https://github.com/wu-yc/scMetabolism}
\item quick start: \url{https://github.com/wu-yc/scMetabolism}
}
}

\concept{Metabolic Analysis}
