% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{OutFromSeuratToOthers}
\alias{OutFromSeuratToOthers}
\title{Convert Seurat Object to Other Format}
\usage{
OutFromSeuratToOthers(obj, format, outfile = NULL)
}
\arguments{
\item{csv}{The header must be "project,path"}
}
\value{
Returns List consisted of multiple Seurat objects
}
\description{
convert a seurat object to other format object
}
\concept{utility}
