% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{PlotCellMarkerExpression}
\alias{PlotCellMarkerExpression}
\title{Plot Cell Marker Expression}
\usage{
PlotCellMarkerExpression(input, genes, kw = "celltype")
}
\arguments{
\item{input}{seurat object}

\item{genes}{seurat markergenes vector}
}
\value{
Returns plot object
}
\description{
bar plot of cell ratios
}
