% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{PlotCellRatio}
\alias{PlotCellRatio}
\title{Plot Cell Type Ratio}
\usage{
PlotCellRatio(input, sample = "orig.ident", celltype = "seurat_clusters")
}
\arguments{
\item{input}{seurat object}

\item{p}{seurat plot}
}
\value{
Returns plot object
}
\description{
bar plot of cell ratios
}
\concept{utility}
