% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{qc}
\alias{qc}
\alias{group_qc}
\title{Quality Control}
\usage{
qc(
  indir,
  outdir,
  project,
  max.counts,
  min.counts,
  max.genes,
  min.genes,
  min.cells,
  max.mt,
  max.hb,
  genecol = 2
)

group_qc(
  csv,
  outdir,
  project,
  max.counts,
  min.counts,
  max.genes,
  min.genes,
  min.cells,
  max.mt,
  max.hb,
  genecol = 2
)
}
\description{
from cellranger matrix to cluster, including normalization, scaling,
dimension reduction, clustering
}
