% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ReadtoSeuratObject}
\alias{ReadtoSeuratObject}
\title{Read files to Seurat Object}
\usage{
ReadtoSeuratObject(infile)
}
\arguments{
\item{csv}{The header must be "project,path"}
}
\value{
Returns List consisted of multiple Seurat objects
}
\description{
Merge data included in the csvfile to one list
}
\concept{utility}
