% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scicolors}
\alias{scicolors}
\title{Science Colors}
\usage{
scicolors(x, journal = "nature")
}
\arguments{
\item{x}{length}
}
\value{
Returns vector of fixed length colors
}
\description{
Science Color Platte
}
\concept{utility}
