% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/trajectory.R
\name{Trajectory}
\alias{Trajectory}
\alias{Trajectory.Monocole3}
\alias{Trajectory.scVelo}
\title{Trajectory Analysis}
\usage{
Trajectory(csv, outdir, used = "Monocole3")

\method{Trajectory}{Monocole3}(
  input,
  outdir,
  used,
  reduction = "UMAP",
  root.key = NULL,
  root.value = NULL
)

\method{Trajectory}{scVelo}(input, outdir, used)
}
\arguments{
\item{outdir}{The ouput directory for rds and plot}

\item{used}{Methods for  trajectories: Monocole3, scVelo}

\item{root.key}{Key used to find root}

\item{root.value}{Value of root key}

\item{object}{The header is "project,path", the path must included matrix.mtx, genes.tsv, bardcodes.tsv}

\item{reduction_method}{Currently "UMAP", "tSNE", "PCA", "LSI", and "Aligned" are supported.}
}
\value{
Return a combined object
}
\description{
Trajectory Analysis
}
\section{Monocole3}{

\itemize{
\item source code: \url{https://github.com/cole-trapnell-lab/monocle3}
\item quick start: \url{https://cole-trapnell-lab.github.io/monocle3/docs/trajectories/}
\item seuratwrapper: \url{http://htmlpreview.github.io/?https://github.com/satijalab/seurat-wrappers/blob/master/docs/monocle3.html}
}
}

\section{scVelo}{

\itemize{
\item source code: \url{https://github.com/dviraran/SingleR}
\item quick start: \url{https://bioconductor.org/packages/devel/bioc/vignettes/SingleR/inst/doc/SingleR.html}
}
}

\concept{trajectory}
