% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{clustering}
\alias{clustering}
\title{Clustering on seurat object}
\usage{
clustering(
  input,
  outdir,
  project,
  nfeatures = 2000,
  plot.features = c("nFeature_RNA", "percent.mt", "percent.rb"),
  reduction = "pca",
  dims = 30,
  k = 20,
  resolution = 0.8
)
}
\arguments{
\item{input}{Input seurat object}

\item{outdir}{Output folder}

\item{project}{Project name}

\item{nfeatures}{Number of variable features to used}

\item{plot.features}{Features to plot on UMAP, default is c("nFeature_RNA", "percent.mt", "percent.rb")}

\item{reduction}{Reduction method, current support is pca, harmony, iNMF (liger)}

\item{resolution}{Resolution to use for clustering}

\item{dim}{Dimensions to use for clustering}
}
\description{
Clustering on seurat object
}
