% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dedoublet.R
\name{group_remove_doublet}
\alias{group_remove_doublet}
\title{Remove doublet on grouped seurat object}
\usage{
group_remove_doublet(
  input,
  outdir,
  project,
  dims,
  nfeatures = 2000,
  resolution = 2,
  cores = 10
)
}
\arguments{
\item{input}{Input seurat object}

\item{outdir}{Output folder}

\item{project}{Project name}

\item{dims}{Dimensions to use for clustering}

\item{nfeatures}{Number of variable features to used}
}
\description{
Remove doublet on grouped seurat object
}
