% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/integration.R
\name{Integration}
\alias{Integration}
\alias{Integration.SeuratCCA}
\alias{Integration.SeuratLargeData}
\alias{Integration.SCTransform}
\alias{Integration.Harmony}
\alias{Integration.Liger}
\title{Project Data Integration}
\usage{
Integration(
  csv,
  outdir,
  used = "SeuratCCA",
  mincell = 3,
  minrna = 200,
  maxrna = 2500,
  maxmt = 5
)

\method{Integration}{SeuratCCA}(
  csv,
  outdir,
  used,
  mincell = 3,
  minrna = 200,
  maxrna = 2500,
  maxmt = 5
)

\method{Integration}{SeuratLargeData}(
  csv,
  outdir,
  used,
  mincell = 3,
  minrna = 200,
  maxrna = 2500,
  maxmt = 5,
  reference = c(1, 2)
)

\method{Integration}{SCTransform}(
  csv,
  outdir,
  used,
  mincell = 3,
  minrna = 200,
  maxrna = 2500,
  maxmt = 5
)

\method{Integration}{Harmony}(
  csv,
  outdir,
  used,
  mincell = 3,
  minrna = 200,
  maxrna = 2500,
  maxmt = 5
)

\method{Integration}{Liger}(
  csv,
  outdir,
  used,
  mincell = 3,
  minrna = 200,
  maxrna = 2500,
  maxmt = 5
)
}
\arguments{
\item{csv}{The header is "project,path", the path must included matrix.mtx, genes.tsv, bardcodes.tsv}

\item{outdir}{The ouput directory for rds and plot}

\item{used}{Methods for integration: SeuratCCA, SCTransform, Harmony, Liger, default is SeuratCCA}

\item{reference}{Sample IDs to used as integration}
}
\value{
Return a combined object
}
\description{
Project Data Integration
}
\section{Seurat CCA}{

\itemize{
\item source code: \url{https://github.com/satijalab/seurat}
\item quick start: \url{https://satijalab.org/seurat/articles/integration_introduction.html#performing-integration-on-datasets-normalized-with-sctransform-1}
}
}

\section{Seurat Large Data}{

For very large datasets, Seruat provide two options that can improve efficiency and runtimes:
\itemize{
\item Reciprocal PCA (RPCA): faster and more conservative
\item Reference-based integration
}
\itemize{
\item quick start: \url{https://satijalab.org/seurat/articles/integration_large_datasets.html}
\item RPCA: \url{https://satijalab.org/seurat/articles/integration_rpca.html}
}
}

\section{SCTransform Integration}{

\itemize{
\item source code: \url{https://github.com/satijalab/seurat}
\item quick start: \url{https://satijalab.org/seurat/articles/integration_introduction.html#performing-integration-on-datasets-normalized-with-sctransform-1}
}
}

\section{Integration Using Harmony}{

\itemize{
\item source code: \url{https://github.com/immunogenomics/harmony}
\item quick start: \url{https://portals.broadinstitute.org/harmony/articles/quickstart.html}
}
}

\section{Integration Using Liger}{

\itemize{
\item source code: \url{https://github.com/welch-lab/liger}
\item quick start: \url{https://htmlpreview.github.io/?https://github.com/satijalab/seurat.wrappers/blob/master/docs/liger.html}
}
}

\concept{integration}
