% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dedoublet.R
\name{remove_doublet}
\alias{remove_doublet}
\title{Remove doublet on seurat object}
\usage{
remove_doublet(
  input,
  outdir,
  project,
  dims,
  nfeatures = 2000,
  resolution = 1,
  cores = 10
)
}
\arguments{
\item{input}{Input seurat object}

\item{outdir}{Output folder}

\item{project}{Project name}

\item{dims}{Dimensions to use for clustering}

\item{nfeatures}{Number of variable features to used}

\item{cores}{Cores to compute}
}
\description{
Remove doublet on seurat object
}
