% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dedoublet.R
\name{remove_doublet}
\alias{remove_doublet}
\title{Remove doublet on seurat object}
\usage{
remove_doublet(
  input,
  outdir,
  project,
  nfeatures = 2000,
  dims = 30,
  anno.used = NULL
)
}
\arguments{
\item{input}{Input seurat object}

\item{outdir}{Output folder}

\item{project}{Project name}

\item{nfeatures}{Number of variable features to used}

\item{dims}{Dimensions to use for clustering}

\item{anno.used}{Used annotation, SingleR or scCATCH, default is NULL}
}
\description{
Remove doublet on seurat object
}
